chore: compare storage script
