+++ eslint.config.mjs
@@ -381,7 +381,6 @@ export default [
     files: [
       "packages/injected/src/**/*.ts",
       "packages/playwright-core/src/server/pageBinding.ts",
-      "packages/playwright-core/src/server/storageScript.ts",
     ],
     languageOptions: languageOptionsWithTsConfig,
     rules: {
